Extract DownloadCache from Rest class
